requirements.txt
Flask==3.1.1
injector==0.22.0
python-dotenv==1.1.1
Requests==2.32.4